renamed react files to fit .jsx format
